+++ umbral/params.py
@@ -3,8 +3,8 @@ from cryptography.hazmat.primitives.asymmetric import ec
 
 class UmbralParameters(object):
     def __init__(self, curve: ec.EllipticCurve):
-        from umbral.point import Point
-        from umbral.utils import unsafe_hash_to_point
+        from umbral.point import Point, unsafe_hash_to_point
+        from umbral.utils import get_curve_keysize_bytes
 
         self.curve = curve
 
@@ -14,7 +14,7 @@ class UmbralParameters(object):
         g_bytes = self.g.to_bytes(is_compressed=True)
 
         self.CURVE_MINVAL_SHA512 = (1 << 512) % int(self.order)
-        self.CURVE_KEY_SIZE_BYTES = self.curve.key_size // 8
+        self.CURVE_KEY_SIZE_BYTES = get_curve_keysize_bytes(self.curve)
 
         domain_seed = b'NuCypherKMS/UmbralParameters/'
         self.h = unsafe_hash_to_point(self, g_bytes, domain_seed + b'h')
